+++ obj/src/persistence/SqlitePersistence.js
@@ -88,7 +88,7 @@ class SqlitePersistence {
      * @param tableName    (optional) a table name.
      */
     constructor(tableName) {
-        this._autoObjects = [];
+        this._schemaStatements = [];
         /**
          * The dependency resolver.
          */
@@ -176,11 +176,31 @@ class SqlitePersistence {
         this.autoCreateObject(builder);
     }
     /**
-     * Adds index definition to create it on opening
-     * @param dmlStatement DML statement to autocreate database object
+     * Adds a statement to schema definition.
+     * This is a deprecated method. Use ensureSchema instead.
+     * @param schemaStatement a statement to be added to the schema
      */
-    autoCreateObject(dmlStatement) {
-        this._autoObjects.push(dmlStatement);
+    autoCreateObject(schemaStatement) {
+        this.ensureSchema(schemaStatement);
+    }
+    /**
+     * Adds a statement to schema definition
+     * @param schemaStatement a statement to be added to the schema
+     */
+    ensureSchema(schemaStatement) {
+        this._schemaStatements.push(schemaStatement);
+    }
+    /**
+     * Clears all auto-created objects
+     */
+    clearSchema() {
+        this._schemaStatements = [];
+    }
+    /**
+     * Defines database schema via auto create objects or convenience methods.
+     */
+    defineSchema() {
+        // Todo: override in chile classes
     }
     /**
      * Converts object value from internal to public format.
@@ -245,8 +265,10 @@ class SqlitePersistence {
             else {
                 this._client = this._connection.getConnection();
                 this._databaseName = this._connection.getDatabaseName();
+                // Define database schema
+                this.defineSchema();
                 // Recreate objects
-                this.autoCreateObjects(correlationId, (err) => {
+                this.createSchema(correlationId, (err) => {
                     if (err) {
                         this._client == null;
                         err = new pip_services3_commons_node_4.ConnectionException(correlationId, "CONNECT_FAILED", "Connection to sqlite failed").withCause(err);
@@ -318,13 +340,13 @@ class SqlitePersistence {
                 callback(err);
         });
     }
-    autoCreateObjects(correlationId, callback) {
-        if (this._autoObjects == null || this._autoObjects.length == 0) {
+    createSchema(correlationId, callback) {
+        if (this._schemaStatements == null || this._schemaStatements.length == 0) {
             callback(null);
             return null;
         }
         // Check if table exist to determine weither to auto create objects
-        let query = "SELECT * FROM '" + this._tableName + "' LIMIT 1";
+        let query = "SELECT * FROM " + this.quoteIdentifier(this._tableName) + " LIMIT 1";
         this._client.get(query, (err) => {
             // If table already exists then exit
             if (err == null) {
@@ -337,7 +359,7 @@ class SqlitePersistence {
             }
             this._logger.debug(correlationId, 'Table ' + this._tableName + ' does not exist. Creating database objects...');
             // Run all DML commands
-            async.eachSeries(this._autoObjects, (dml, callback) => {
+            async.eachSeries(this._schemaStatements, (dml, callback) => {
                 this._client.exec(dml, (err) => {
                     if (err) {
                         this._logger.error(correlationId, err, 'Failed to autocreate database object');