+++ test/persistence/DummyJsonSqlitePersistence.ts
@@ -12,8 +12,12 @@ export class DummyJsonSqlitePersistence
 {
     public constructor() {
         super('dummies_json');
+    }
+
+    protected defineSchema(): void {
+        this.clearSchema();
         this.ensureTable();
-        this.autoCreateObject("CREATE UNIQUE INDEX IF NOT EXISTS dummies_json_key ON dummies_json (JSON_EXTRACT(data, '$.key'))");
+        this.ensureSchema("CREATE UNIQUE INDEX IF NOT EXISTS \"" + this._tableName + "_json_key\" ON dummies_json (JSON_EXTRACT(data, '$.key'))");
     }
 
     public getPageByFilter(correlationId: string, filter: FilterParams, paging: PagingParams,